chore: move `qopen.cpp` to `src/game/qopen.cpp`
