+++ src/game/qopen.cpp
