+++ src/pic/pic8.cpp
@@ -3,7 +3,7 @@
 #include "pic/pic8.h"
 #include "platform/implementation.h"
 #include "platform/utils.h"
-#include "qopen.h"
+#include "game/qopen.h"
 #include <algorithm>
 #include <cstring>
 #include <climits>
